Declare the koji-build script
