+++ scripts/koji/.gitignore
@@ -1,1 +1,5 @@
 __pycache__
+
+# Distribution / packaging
+dist/
+*.egg-info/
