+++ scripts/koji/pyproject.toml
@@ -1,13 +1,16 @@
 [project]
 name = "koji-utils"
-version = "0.1.0"
 description = "Koji utils"
 readme = "README.md"
 requires-python = "~=3.11"
+dynamic = ["version"]
 dependencies = [
   "koji",
   "specfile",
 ]
+
+[project.scripts]
+koji-build = "koji_utils.koji_build:main"
 
 [dependency-groups]
 dev = [
@@ -19,6 +22,13 @@ dev = [
   "typing-extensions",
   "flake8-pyproject",
 ]
+
+[build-system]
+requires = ["setuptools >= 77.0.3", "setuptools-scm>=8", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools_scm]
+root = "../.."
 
 [tool.pyright]
 typeCheckingMode = "standard"
